Fix another Windows warning

+++ clients/drcachesim/tracer/instru_offline.cpp
@@ -95,7 +95,7 @@ offline_instru_t::offline_instru_t(
     uint64 max_bb_instrs;
     if (!dr_get_integer_option("max_bb_instrs", &max_bb_instrs))
         max_bb_instrs = 256; /* current default */
-    max_block_encoding_size_ = max_bb_instrs * MAX_INSTR_LENGTH;
+    max_block_encoding_size_ = static_cast<int>(max_bb_instrs) * MAX_INSTR_LENGTH;
     encoding_lock_ = dr_mutex_create();
     encoding_buf_sz_ = ALIGN_FORWARD(max_block_encoding_size_ * 10, dr_page_size());
     encoding_buf_start_ = reinterpret_cast<byte *>(
